refactor ts compiler writer
