+++ compile/ts/compiler.go
@@ -1,7 +1,6 @@
 package tscode
 
 import (
-	"bytes"
 	"fmt"
 	"reflect"
 	"strings"
@@ -13,8 +12,7 @@ import (
 
 // Compiler translates a Mochi AST into TypeScript source code that can be run with Deno.
 type Compiler struct {
-	buf          bytes.Buffer
-	indent       int
+	emitter
 	helpers      map[string]bool
 	env          *types.Env
 	structs      map[string]bool
@@ -25,6 +23,7 @@ type Compiler struct {
 // New creates a new TypeScript compiler instance.
 func New(env *types.Env) *Compiler {
 	return &Compiler{
+		emitter: emitter{},
 		helpers: make(map[string]bool),
 		env:     env,
 		structs: make(map[string]bool),
@@ -71,6 +70,8 @@ func stmtHasStream(s *parser.Statement) bool {
 
 // Compile generates TypeScript source code for the given program.
 func (c *Compiler) Compile(prog *parser.Program) ([]byte, error) {
+	c.buf.Reset()
+	c.indent = 0
 	needsAsync := containsStreamCode(prog.Statements)
 
 	c.writeln("// Generated by Mochi TypeScript compiler")
@@ -1401,30 +1402,6 @@ func (c *Compiler) compileLiteral(l *parser.Literal) (string, error) {
 }
 
 // --- Helpers ---
-
-func (c *Compiler) writeln(s string) {
-	c.writeIndent()
-	c.buf.WriteString(s)
-	c.buf.WriteByte('\n')
-}
-
-func (c *Compiler) writeIndent() {
-	for i := 0; i < c.indent; i++ {
-		c.buf.WriteByte('\t')
-	}
-}
-
-func indentBlock(s string, depth int) string {
-	if s == "" {
-		return s
-	}
-	prefix := strings.Repeat("\t", depth)
-	lines := strings.Split(strings.TrimRight(s, "\n"), "\n")
-	for i, line := range lines {
-		lines[i] = prefix + line
-	}
-	return strings.Join(lines, "\n") + "\n"
-}
 
 func (c *Compiler) compileStructType(st types.StructType) {
 	name := sanitizeName(st.Name)
